Extracted user info with middleware. Added extra tests

+++ controllers/blogs.js
@@ -1,23 +1,20 @@
 const blogRouter = require('express').Router();
-const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
-const { tokenExtractor } = require('../utils/middleware');
 
 blogRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({}).populate('user', { username: 1 });
   res.json(blogs);
 });
 
-blogRouter.post('/', tokenExtractor, async (req, res) => {
-  if (!req.token) return res.status(401).send({ error: 'Missing token' });
+blogRouter.post('/', async (req, res) => {
+  if (!req.token || !req.user) return res.status(401).end();
 
-  const tokenUser = jwt.verify(req.token, process.env.SECRET);
-
-  const user = await User.findById(tokenUser.id);
+  const user = await User.findById(req.user.id);
   if (!user) return res.status(404).send({ error: 'User not found' });
 
-  req.body.user = user.id;
+  // eslint-disable-next-line no-underscore-dangle
+  req.body.user = user._id;
   const blog = new Blog(req.body);
   const savedBlog = await blog.save();
 
@@ -27,14 +24,12 @@ blogRouter.post('/', tokenExtractor, async (req, res) => {
   return res.status(201).json(savedBlog);
 });
 
-blogRouter.delete('/:id', tokenExtractor, async (req, res) => {
-  if (!req.token) return res.status(401).end();
-
-  const tokenUser = jwt.verify(req.token, process.env.SECRET);
+blogRouter.delete('/:id', async (req, res) => {
+  if (!req.token || !req.user) return res.status(401).end();
   const blog = await Blog.findById(req.params.id);
   if (!blog) return res.status(404).end();
 
-  if (blog.user.toString() !== tokenUser.id) return res.status(400).send({ error: 'Blog owner does not match token owner' });
+  if (blog.user.toString() !== req.user.id) return res.status(400).send({ error: 'Blog owner does not match token owner' });
   Blog.deleteOne(blog);
 
   return res.status(204).end();
